+++ src/components/features/InfoPoper.tsx
@@ -1,0 +1,45 @@
+import * as React from 'react'
+import Box from '@mui/material/Box'
+import Popper, { PopperPlacementType } from '@mui/material/Popper'
+import Grid from '@mui/material/Grid'
+import Button from '@mui/material/Button'
+import Fade from '@mui/material/Fade'
+import Paper from '@mui/material/Paper'
+import InfoIcon from '@mui/icons-material/Info'
+
+interface InfoPopperProps {
+  content: JSX.Element
+}
+
+export default function InfoPopper({ content }: InfoPopperProps) {
+  const [anchorEl, setAnchorEl] = React.useState<HTMLButtonElement | null>(null)
+  const [open, setOpen] = React.useState(false)
+  const [placement, setPlacement] = React.useState<PopperPlacementType>()
+
+  const handleClick =
+    (newPlacement: PopperPlacementType) => (event: React.MouseEvent<HTMLButtonElement>) => {
+      setAnchorEl(event.currentTarget)
+      setOpen((prev) => placement !== newPlacement || !prev)
+      setPlacement(newPlacement)
+    }
+
+  return (
+    <Box>
+      <Popper open={open} anchorEl={anchorEl} placement={placement} transition>
+        {({ TransitionProps }) => (
+          <Fade {...TransitionProps} timeout={350}>
+            <Paper sx={{ width: '200px' }}>{content}</Paper>
+          </Fade>
+        )}
+      </Popper>
+
+      <Grid container justifyContent="center">
+        <Grid item>
+          <Button onClick={handleClick('bottom-start')}>
+            <InfoIcon />
+          </Button>
+        </Grid>
+      </Grid>
+    </Box>
+  )
+}
